- add grayscale for iron material prefabs - fix colliders for ships by removing them for rigidbody and   BlockingCollider. May need to add a few specific omissions, but the   main issue was ignoring piece caused the boat to phase through piece   objects
