+++ src/ValheimRAFT/AssetBundle/AssetBundle.manifest
@@ -1,5 +1,5 @@
 ManifestFileVersion: 0
-CRC: 259722753
+CRC: 3467557213
 AssetBundleManifest:
   AssetBundleInfos:
     Info_0:
