+++ src/ValheimRAFT/AssetBundle/valheim-vehicles.manifest
@@ -1,15 +1,15 @@
 ManifestFileVersion: 0
-CRC: 4218703176
+CRC: 4024397762
 Hashes:
   AssetFileHash:
     serializedVersion: 2
-    Hash: a1daca586ba926287b0fdbd8313c2a1c
+    Hash: 2a5637638dfdc0fbcdc38913dab30336
   TypeTreeHash:
     serializedVersion: 2
     Hash: 9de5c4e5f8d12d9935b13bf93fed7941
   IncrementalBuildHash:
     serializedVersion: 2
-    Hash: a1daca586ba926287b0fdbd8313c2a1c
+    Hash: 2a5637638dfdc0fbcdc38913dab30336
 HashAppended: 0
 ClassTypes:
 - Class: 1
