+++ src/ValheimRAFT/ValheimRAFT.Unity/Assets/ValheimVehicles/Material/iron_metal_custom.mat
@@ -1,0 +1,151 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!21 &2100000
+Material:
+  serializedVersion: 8
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: iron_metal_custom
+  m_Shader: {fileID: 4800000, guid: 906ebf25ce541ed4ab17c721e09594b9, type: 3}
+  m_Parent: {fileID: 0}
+  m_ModifiedSerializedProperties: 0
+  m_ValidKeywords:
+  - _DETAIL_MULX2
+  - _EMISSION
+  - _NORMALMAP
+  m_InvalidKeywords:
+  - _TWOSIDEDNORMALS_ON
+  m_LightmapFlags: 2
+  m_EnableInstancingVariants: 1
+  m_DoubleSidedGI: 1
+  m_CustomRenderQueue: -1
+  stringTagMap: {}
+  disabledShaderPasses: []
+  m_LockedProperties: 
+  m_SavedProperties:
+    serializedVersion: 3
+    m_TexEnvs:
+    - _BumpMap:
+        m_Texture: {fileID: 2800000, guid: 6a73eed2143f04947aaeaaf1070fb034, type: 3}
+        m_Scale: {x: 5, y: 5}
+        m_Offset: {x: 0, y: 0}
+    - _DetailAlbedoMap:
+        m_Texture: {fileID: 0}
+        m_Scale: {x: 20, y: 20}
+        m_Offset: {x: 0, y: 0}
+    - _DetailMask:
+        m_Texture: {fileID: 0}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _DetailNormalMap:
+        m_Texture: {fileID: 2800000, guid: 6a73eed2143f04947aaeaaf1070fb034, type: 3}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _EmissionMap:
+        m_Texture: {fileID: 0}
+        m_Scale: {x: 0.3, y: 0.3}
+        m_Offset: {x: 0, y: 0}
+    - _EmissiveTex:
+        m_Texture: {fileID: 0}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _LogoNormal:
+        m_Texture: {fileID: 2800000, guid: 6a73eed2143f04947aaeaaf1070fb034, type: 3}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _LogoTex:
+        m_Texture: {fileID: 2800000, guid: dbff036d19c6854489c244673c4ee246, type: 3}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _MainColorNormal:
+        m_Texture: {fileID: 2800000, guid: 6a73eed2143f04947aaeaaf1070fb034, type: 3}
+        m_Scale: {x: 0.1, y: 0.1}
+        m_Offset: {x: 0, y: 0}
+    - _MainColorNormalMap:
+        m_Texture: {fileID: 0}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _MainTex:
+        m_Texture: {fileID: 2800000, guid: 4352d9e89bfd7f944a3b241d5fec6c5c, type: 3}
+        m_Scale: {x: 0.3, y: 0.3}
+        m_Offset: {x: 0, y: 0}
+    - _MetallicGlossMap:
+        m_Texture: {fileID: 0}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _MossTex:
+        m_Texture: {fileID: 0}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _OcclusionMap:
+        m_Texture: {fileID: 0}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _ParallaxMap:
+        m_Texture: {fileID: 0}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _PatternNormal:
+        m_Texture: {fileID: 2800000, guid: a8473ccfbdc97294cb35f727bdf15949, type: 3}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _PatternTex:
+        m_Texture: {fileID: 2800000, guid: dbff036d19c6854489c244673c4ee246, type: 3}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    - _SpecGlossMap:
+        m_Texture: {fileID: 0}
+        m_Scale: {x: 1, y: 1}
+        m_Offset: {x: 0, y: 0}
+    m_Ints: []
+    m_Floats:
+    - _AddRain: 0
+    - _AddSnow: 0
+    - _BumpScale: 2
+    - _CamCull: 0
+    - _Cull: 0
+    - _Cutoff: 0
+    - _DetailNormalMapScale: 1
+    - _DstBlend: 0
+    - _GlossMapScale: 1
+    - _Glossiness: 0.527
+    - _GlossyReflections: 1
+    - _Height: 0
+    - _LogoRotation: 0.468
+    - _Metallic: 1
+    - _MistAlpha: 1
+    - _Mode: 0
+    - _MossAlpha: 1
+    - _MossBlend: 6.65
+    - _MossNormal: 0
+    - _MossTransition: 0.1
+    - _OcclusionStrength: 1
+    - _Parallax: 0.02
+    - _PatternRotation: 1
+    - _PushClothMode: 0
+    - _PushDistance: 0
+    - _RippleDeadzoneMax: 3.08
+    - _RippleDeadzoneMin: 3.93
+    - _RippleDistance: 0.5
+    - _RippleSpeed: 100
+    - _SmoothnessTextureChannel: 0
+    - _SpecularHighlights: 1
+    - _SphereNormals: 0
+    - _SphereOffset: 0
+    - _SrcBlend: 1
+    - _SwayDistance: 0.5
+    - _SwaySpeed: 0
+    - _TwoSidedNormals: 1
+    - _UVSec: 0
+    - _Wet: 0.428
+    - _ZWrite: 1
+    m_Colors:
+    - _Color: {r: 1, g: 1, b: 1, a: 1}
+    - _EmissionColor: {r: 0.03773582, g: 0.03773582, b: 0.03773582, a: 1}
+    - _LogoColor: {r: 1, g: 0.6745283, b: 0.3726415, a: 1}
+    - _MainColor: {r: 0.72156864, g: 0.4509804, b: 0.2, a: 1}
+    - _PatternColor: {r: 0.5471698, g: 0.2092195, b: 0, a: 1}
+    - _SpecColor: {r: 0.2, g: 0.2, b: 0.2, a: 1}
+  m_BuildTextureStacks: []
