+++ src/ValheimRAFT/ValheimRAFT.Unity/Assets/ValheimVehicles/Prefabs/vehicle_ship_pieces.prefab
@@ -10,7 +10,7 @@ GameObject:
   m_Component:
   - component: {fileID: 370483867695338920}
   - component: {fileID: 6223762886225816425}
-  m_Layer: 0
+  m_Layer: 29
   m_Name: vehicle_ship_pieces
   m_TagString: Untagged
   m_Icon: {fileID: 0}
@@ -51,11 +51,11 @@ Rigidbody:
     m_Bits: 0
   m_ExcludeLayers:
     serializedVersion: 2
-    m_Bits: 0
+    m_Bits: 58925056
   m_ImplicitCom: 1
   m_ImplicitTensor: 1
   m_UseGravity: 0
   m_IsKinematic: 1
-  m_Interpolate: 2
+  m_Interpolate: 1
   m_Constraints: 0
-  m_CollisionDetection: 0
+  m_CollisionDetection: 1
