+++ src/ValheimVehicles/ValheimVehicles.Vehicles/Components/PlayerSpawnController.cs
@@ -8,6 +8,7 @@ using UnityEngine;
 using UnityEngine.UI;
 using ValheimRAFT.Util;
 using ValheimVehicles.Prefabs;
+using ValheimVehicles.ValheimVehicles.DynamicLocations;
 using ValheimVehicles.Vehicles;
 using Logger = Jotunn.Logger;
 
@@ -25,222 +26,8 @@ namespace ValheimVehicles.Vehicles.Components;
 public class PlayerSpawnController : MonoBehaviour
 {
   // spawn (Beds)
-
-  public static class DynamicLocations
-  {
-    private static string PluginPrefix = "valheim_vehicles";
-    private const string DynamicPrefix = "Dynamic";
-    private const string SpawnZdo = "SpawnZdo";
-    private const string LogoutParentZdoOffset = "LogoutParentZdoOffset";
-    private const string LogoutParentZdo = "LogoutParentZdo";
-
-    public static string GetPluginPrefix()
-    {
-      return PluginPrefix;
-    }
-
-    public static string GetFullPrefix()
-    {
-      return $"{PluginPrefix}_{DynamicPrefix}";
-    }
-
-    /// <summary>
-    /// must be guarded
-    /// </summary>
-    private static long WorldUID => ZNet.instance?.GetWorldUID() ?? 0;
-
-    public static bool IsValid(string key) => key.StartsWith(PluginPrefix);
-
-    public static string GetLogoutZdoOffsetKey() =>
-      !ZNet.instance ? "" : $"{GetFullPrefix()}_{LogoutParentZdoOffset}_{WorldUID}";
-
-    public static string GetLogoutZdoKey() => !ZNet.instance
-      ? ""
-      : $"{GetFullPrefix()}_{LogoutParentZdo}_{WorldUID}";
-
-    public static string GetSpawnZdoOffsetKey() =>
-      !ZNet.instance ? "" : $"{GetFullPrefix()}_{LogoutParentZdoOffset}_{WorldUID}";
-
-    public static string GetSpawnZdoKey() =>
-      !ZNet.instance ? "" : $"{GetFullPrefix()}_{SpawnZdo}_{WorldUID}";
-
-    private static string ZDOIDToString(ZDOID zdoid)
-    {
-      var userId = zdoid.UserID;
-      var id = zdoid.ID;
-      return $"{userId},{id}";
-    }
-
-    private static ZDOID? StringToZDOID(string zdoidString)
-    {
-      var zdoIdStringArray = zdoidString.Split(',');
-      if (zdoIdStringArray.Length != 2) return null;
-
-      long.TryParse(zdoIdStringArray[0], out var userId);
-      uint.TryParse(zdoIdStringArray[1], out var objectId);
-
-      if (userId == 0 || objectId == 0)
-      {
-        Logger.LogDebug("failed to parse to ZDOID");
-        return null;
-      }
-
-      var zdoID = new ZDOID(userId, objectId);
-      return zdoID;
-    }
-
-    private static string Vector3ToString(Vector3 val)
-    {
-      return $"{val.x},{val.y},{val.z}";
-    }
-
-    private static Vector3? StringToVector3(string val)
-    {
-      var vector3StringArray = val.Split(',');
-      if (vector3StringArray.Length != 3) return null;
-
-      float.TryParse(vector3StringArray[0], out var x);
-      float.TryParse(vector3StringArray[1], out var y);
-      float.TryParse(vector3StringArray[2], out var z);
-      var vector = new Vector3(x, y, z);
-      return vector;
-    }
-
-    public static ZDOID? GetLogoutZdo(Player player)
-    {
-      if (!player) return null;
-      if (!player.m_customData.TryGetValue(GetLogoutZdoKey(), out var logoutZdoString))
-      {
-        return null;
-      }
-
-      var zdoid = StringToZDOID(logoutZdoString);
-      return zdoid;
-    }
-
-    public static Vector3 GetLogoutZdoOffset(Player player)
-    {
-      if (!player) return Vector3.zero;
-      if (!player.m_customData.TryGetValue(GetLogoutZdoKey(), out var logoutZdoString))
-      {
-        return Vector3.zero;
-      }
-
-      var offset = StringToVector3(logoutZdoString);
-      return offset ?? Vector3.zero;
-    }
-
-    public static bool SetLogoutZdo(Player player, ZNetView dynamicObj)
-    {
-      if (!ZNet.instance) return false;
-      var spawnPointObjZdo = dynamicObj.GetZDO();
-      if (spawnPointObjZdo == null) return false;
-      player.m_customData[GetLogoutZdoKey()] = ZDOIDToString(spawnPointObjZdo.m_uid);
-      return true;
-    }
-
-    public static bool RemoveLogoutZdo(Player player)
-    {
-      if (!ZNet.instance) return false;
-      player.m_customData.Remove(GetLogoutZdoKey());
-      player.m_customData.Remove(GetLogoutZdoOffsetKey());
-      return true;
-    }
-
-    public static bool RemoveSpawnZdo(Player player)
-    {
-      if (!ZNet.instance) return false;
-      player.m_customData.Remove(GetSpawnZdoKey());
-      player.m_customData.Remove(GetSpawnZdoOffsetKey());
-      return true;
-    }
-
-    public static bool SetLogoutZdoOffset(Player player, Vector3 offset)
-    {
-      if (!player) return false;
-      if (Vector3.zero == offset)
-      {
-        player.m_customData.Remove(GetLogoutZdoOffsetKey());
-        return false;
-      }
-
-      player.m_customData[GetLogoutZdoOffsetKey()] = Vector3ToString(offset);
-      return true;
-    }
-
-    public static bool SetLogoutZdoWithOffset(Player player, ZNetView dynamicObj,
-      Vector3 offset)
-    {
-      if (!SetLogoutZdo(player, dynamicObj)) return false;
-      SetLogoutZdoOffset(player, offset);
-      return true;
-    }
-
-
-    public static ZDOID? GetSpawnZdo(Player player)
-    {
-      if (!player) return null;
-      if (!player.m_customData.TryGetValue(GetSpawnZdoKey(), out var spawnZdoString))
-      {
-        return null;
-      }
-
-      var zdoid = StringToZDOID(spawnZdoString);
-      return zdoid;
-    }
-
-    public static Vector3 GetSpawnZdoOffset(Player player)
-    {
-      if (!player) return Vector3.zero;
-      if (!player.m_customData.TryGetValue(GetSpawnZdoOffsetKey(), out var offsetString))
-      {
-        return Vector3.zero;
-      }
-
-      var offset = StringToVector3(offsetString) ?? Vector3.zero;
-      return offset;
-    }
-
-    public static bool SetSpawnZdo(Player player, ZNetView dynamicObj)
-    {
-      if (!ZNet.instance) return false;
-      var spawnPointObjZdo = dynamicObj.GetZDO();
-      if (spawnPointObjZdo == null) return false;
-      player.m_customData[GetSpawnZdoKey()] = ZDOIDToString(spawnPointObjZdo.m_uid);
-      return true;
-    }
-
-    /// <summary>
-    /// Only sets offset if necessary, otherwise scrubs the data
-    /// </summary>
-    /// <param name="player"></param>
-    /// <param name="offset"></param>
-    /// <returns></returns>
-    public static bool SetSpawnZdoOffset(Player player, Vector3 offset)
-    {
-      if (!player) return false;
-      if (Vector3.zero == offset)
-      {
-        if (player.m_customData.TryGetValue(GetSpawnZdoOffsetKey(), out _))
-        {
-          player.m_customData.Remove(GetSpawnZdoOffsetKey());
-        }
-
-        return false;
-      }
-
-      player.m_customData[GetSpawnZdoKey()] = Vector3ToString(offset);
-      return true;
-    }
-
-    public static bool SetSpawnZdoWithOffset(Player player, ZNetView dynamicObj,
-      Vector3 offset)
-    {
-      if (!SetSpawnZdo(player, dynamicObj)) return false;
-      SetSpawnZdoOffset(player, offset);
-      return true;
-    }
-  }
+  public static bool CanUpdateLogoutPoint = true;
+  public static bool CanRemoveLogoutAfterSync = false;
 
   public ZNetView? NetView;
   public long playerId;
@@ -257,7 +44,7 @@ public class PlayerSpawnController : MonoBehaviour
 
   private void Start()
   {
-    InvokeRepeating(nameof(SyncLogoutPoint), 5f, 5f);
+    // InvokeRepeating(nameof(SyncLogoutPoint), 5f, 5f);
     if (MovePlayerToLoginPoint())
     {
       return;
@@ -266,15 +53,15 @@ public class PlayerSpawnController : MonoBehaviour
     MovePlayerToSpawnPoint();
   }
 
-  private void OnEnable()
-  {
-    InvokeRepeating(nameof(SyncLogoutPoint), 5f, 20);
-  }
-
-  private void OnDisable()
-  {
-    CancelInvoke(nameof(SyncLogoutPoint));
-  }
+  // private void OnEnable()
+  // {
+  //   // InvokeRepeating(nameof(SyncLogoutPoint), 5f, 20);
+  // }
+  //
+  // private void OnDisable()
+  // {
+  //   // CancelInvoke(nameof(SyncLogoutPoint));
+  // }
 
   private void Setup()
   {
@@ -333,7 +120,7 @@ public class PlayerSpawnController : MonoBehaviour
   /// <returns>bool</returns>
   public void SyncLogoutPoint()
   {
-    if (!player || player == null || zdo == null) return;
+    if (!player || player == null || zdo == null || !CanUpdateLogoutPoint) return;
     var bvc = player.GetComponentInParent<BaseVehicleController>();
     if (!bvc)
     {
@@ -407,8 +194,13 @@ public class PlayerSpawnController : MonoBehaviour
 
     StartCoroutine(MovePlayerToZdoWorker(loginZdoid, loginZdoOffset));
 
+    // TODO This must be another coroutine AND only fired after the Move coroutine completes otherwise it WILL break the move coroutine as it deletes the required key.
     // remove logout point after moving the player.
-    DynamicLocations.RemoveLogoutZdo(player);
+    if (CanRemoveLogoutAfterSync)
+    {
+      DynamicLocations.RemoveLogoutZdo(player);
+    }
+
     return true;
   }
 
@@ -419,33 +211,36 @@ public class PlayerSpawnController : MonoBehaviour
     StartCoroutine(MovePlayerToZdoWorker(spawnZdoid, spawnZdoOffset));
   }
 
-  public IEnumerator LoadGameObjectInSectorFromZdo(ZDOID zdoid)
-  {
-    var spawnZdo = ZDOMan.instance.GetZDO(zdoid);
-    ZNetView? zdoNetViewInstance = null;
-    Vector2i zoneId = new();
-
-    player.transform.position = spawnZdo.GetPosition();
-    player.m_nview.GetZDO().SetPosition(spawnZdo.GetPosition());
-    player.m_nview.GetZDO().SetSector(spawnZdo.GetSector());
-    while (zdoNetViewInstance == null)
-    {
-      zoneId = ZoneSystem.instance.GetZone(spawnZdo.GetPosition());
-      ZoneSystem.instance.PokeLocalZone(zoneId);
-      zdoNetViewInstance = ZNetScene.instance.FindInstance(spawnZdo);
-
-      if (zdoNetViewInstance) break;
-      yield return new WaitForFixedUpdate();
-    }
-
-    yield return new WaitUntil(() => ZoneSystem.instance.IsZoneLoaded(zoneId));
-    yield return zdoNetViewInstance;
-  }
+  // public IEnumerator LoadGameObjectInSectorFromZdo(ZDOID zdoid)
+  // {
+  //   var spawnZdo = ZDOMan.instance.GetZDO(zdoid);
+  //   ZNetView? zdoNetViewInstance = null;
+  //   Vector2i zoneId = new();
+  //
+  //   player.transform.position = spawnZdo.GetPosition();
+  //   player.m_nview.GetZDO().SetPosition(spawnZdo.GetPosition());
+  //   player.m_nview.GetZDO().SetSector(spawnZdo.GetSector());
+  //   while (zdoNetViewInstance == null)
+  //   {
+  //     zoneId = ZoneSystem.instance.GetZone(spawnZdo.GetPosition());
+  //     ZoneSystem.instance.PokeLocalZone(zoneId);
+  //     zdoNetViewInstance = ZNetScene.instance.FindInstance(spawnZdo);
+  //
+  //     if (zdoNetViewInstance) break;
+  //     yield return new WaitForFixedUpdate();
+  //   }
+  //
+  //   yield return new WaitUntil(() => ZoneSystem.instance.IsZoneLoaded(zoneId));
+  //   yield return zdoNetViewInstance;
+  // }
 
   private void SyncPlayerPosition(Vector3 newPosition)
   {
     if (ZNetView.m_forceDisableInit || player == null) return;
-    player.m_nview.GetZDO()?.SetPosition(newPosition);
+    var playerZdo = player.m_nview.GetZDO();
+    if (playerZdo == null) return;
+    playerZdo.SetPosition(newPosition);
+    playerZdo.SetSector(ZoneSystem.instance.GetZone(newPosition));
     player.transform.position = newPosition;
   }
 
@@ -464,17 +259,30 @@ public class PlayerSpawnController : MonoBehaviour
       yield break;
     }
 
-    // for nullable using .Value is required
-    var zdoNetviewInstance = LoadGameObjectInSectorFromZdo(zdoid.Value);
-    yield return StartCoroutine(zdoNetviewInstance);
-    // var zdoInstanceObj = LoadGameObjectInSectorFromZdo(spawnZdoid.Value);
-    var isNetView = zdoNetviewInstance.Current?.GetType() == typeof(ZNetView);
-    // if (!isNetView) yield break;
-    var nv = zdoNetviewInstance.Current as ZNetView;
-    BaseVehicleController? bvc = null;
-    if (isNetView && nv != null)
+    // beginning of LoadGameObjectInSector (which cannot be abstracted easily if the return is required)
+    var spawnZdo = ZDOMan.instance.GetZDO((ZDOID)zdoid);
+    ZNetView? zdoNetViewInstance = null;
+    Vector2i zoneId = new();
+
+    while (zdoNetViewInstance == null)
     {
-      bvc = nv
+      zoneId = ZoneSystem.instance.GetZone(spawnZdo.GetPosition());
+      ZoneSystem.instance.PokeLocalZone(zoneId);
+      zdoNetViewInstance = ZNetScene.instance.FindInstance(spawnZdo);
+
+      if (zdoNetViewInstance) break;
+      yield return new WaitForFixedUpdate();
+    }
+
+    yield return new WaitUntil(() => ZoneSystem.instance.IsZoneLoaded(zoneId));
+    // yield return zdoNetViewInstance;
+    // end 
+
+
+    BaseVehicleController? bvc = null;
+    if (zdoNetViewInstance)
+    {
+      bvc = zdoNetViewInstance
         .GetComponentInParent<BaseVehicleController>();
       if (bvc)
       {
@@ -487,26 +295,33 @@ public class PlayerSpawnController : MonoBehaviour
 
     yield return new WaitForFixedUpdate();
 
-    var spawnZdo = ZDOMan.instance.GetZDO((ZDOID)zdoid);
-    var spawnOffset = spawnZdo.GetVec3(VehicleZdoVars.MBPositionHash, Vector3.zero);
+    if (zdoid == ZDOID.None) yield break;
+
+    spawnZdo = ZDOMan.instance.GetZDO((ZDOID)zdoid);
+
+    if (spawnZdo == null) yield break;
+
+    var spawnOffset =
+      spawnZdo?.GetVec3(VehicleZdoVars.MBPositionHash, Vector3.zero) ?? Vector3.zero;
     Logger.LogDebug($"SpawnOffset, {spawnOffset}");
-    nv = ZNetScene.instance.FindInstance(spawnZdo);
+    zdoNetViewInstance = ZNetScene.instance.FindInstance(spawnZdo);
 
     var relativeZdoPos = spawnZdo.GetPosition() + spawnOffset;
-    Logger.LogDebug($"ZDOPos+relative {relativeZdoPos} vs transform.pos {nv?.transform?.position}");
-    if (nv != null)
+    Logger.LogDebug(
+      $"ZDOPos+relative {relativeZdoPos} vs transform.pos {zdoNetViewInstance?.transform?.position}");
+    if (zdoNetViewInstance != null)
     {
-      SyncPlayerPosition(nv.transform.position);
+      SyncPlayerPosition(zdoNetViewInstance.transform.position);
     }
 
     yield return new WaitForSeconds(1);
-    if (nv != null)
+    if (zdoNetViewInstance != null)
     {
       spawnZdo = ZDOMan.instance.GetZDO((ZDOID)zdoid);
       spawnOffset = spawnZdo.GetVec3(VehicleZdoVars.MBPositionHash, Vector3.zero);
       relativeZdoPos = spawnZdo.GetPosition() + spawnOffset;
       Logger.LogDebug(
-        $"ZDOPos+relative {relativeZdoPos} vs transform.pos {nv?.transform?.position}");
+        $"ZDOPos+relative {relativeZdoPos} vs transform.pos {zdoNetViewInstance?.transform?.position}");
       SyncPlayerPosition(relativeZdoPos);
     }
   }
