+++ src/ValheimVehicles/ValheimVehicles.csproj
@@ -40,6 +40,7 @@
         <Compile Include="Patches\Zone_Patch.cs"/>
         <Compile Include="ValheimVehicles.ConsoleCommands\LocateVehicle.cs"/>
         <Compile Include="ValheimVehicles.ConsoleCommands\VehicleCommands.cs"/>
+        <Compile Include="ValheimVehicles.DynamicLocations\DynamicLocations.cs"/>
         <Compile Include="ValheimVehicles.Prefabs\PrefabNames.cs"/>
         <Compile Include="ValheimVehicles.Propulsion\Rudder\SteeringWheelComponent.cs"/>
         <Compile Include="ValheimVehicles.Propulsion\Sail\SailAreaForce.cs"/>
@@ -52,7 +53,7 @@
     <ItemGroup>
         <Folder Include="ValheimVehicles.Propulsion\Sailing\"/>
         <Folder Include="ValheimVehicles.Utils\"/>
-        <Folder Include="ValheimVehicles.Vehicles\" />
+        <Folder Include="ValheimVehicles.Vehicles\"/>
     </ItemGroup>
     <ItemGroup>
         <ProjectReference Include="..\SentryUnityWrapper\SentryUnityWrapper.csproj"/>
